Fix : 기존에 Seat 테이블을 전부 Select해서 찾아오던 부분을 Bus 테이블을 사용하도록 변경
Seat는 노선하나에 상당히 많은 insert가 발생하는 반면, Bus는 수가 한정되어 있기 때문에 Bus를 통해 전체 좌석수를 구하도록 변경

+++ BE/src/main/java/com/lisade/togeduck/dto/response/RouteDetailDao.java
@@ -17,7 +17,7 @@ public class RouteDetailDao {
     String source;
     String destination;
     LocalTime arrivalAt;
-    Long totalSeats;
+    Integer totalSeats;
     Long reservedSeats;
     Integer cost;
